Refactored name and location of page components.
